+++ src/external_pkgs/hector-quadrotor-noetic/hector_localization/hector_pose_estimation_core/src/pose_estimation.cpp
@@ -50,7 +50,7 @@ PoseEstimation::PoseEstimation(const SystemPtr& system, const StatePtr& state)
   if (!the_instance) the_instance = this;
   if (system) addSystem(system);
 
-  world_frame_ = "/world";
+  world_frame_ = "world";
   nav_frame_ = "nav";
   base_frame_ = "base_link";
   stabilized_frame_ = "base_stabilized";